+++ SpeedyWeather/src/variables/variables.jl
@@ -368,12 +368,26 @@ function allocate(group, model, fuse_parents = Dict{Tuple{Symbol, Symbol}, Named
 
     # variables without namespace identified by empty symbol Symbol() go directly into the main NamedTuple
     # that way we have variables.prognostic.vorticity skipping the namespace between prognostic and vor
+<<<<<<< mg/gpu-rework-test
     nt1 = haskey(group, Symbol()) ? _allocate_namespace(group[Symbol()], model, fuse_parents) : NamedTuple()
+=======
+    nt1 = NamedTuple{Tuple(map(v -> v.name, group[Symbol()]))}(Tuple(map(var -> allocate(var, model), group[Symbol()])))
+>>>>>>> main
 
     # other variables grouped by namespace
     # e.g. variables.prognostic.ocean.sea_surface_temperature, variables.prognostic.land.soil_moisture, etc.
     nt2 = NamedTuple{namespaces}(
+<<<<<<< mg/gpu-rework-test
         Tuple(map(ns -> _allocate_namespace(group[ns], model, fuse_parents), namespaces))
+=======
+        Tuple(
+            map(
+                ns ->
+                NamedTuple{Tuple(map(v -> v.name, group[ns]))}(Tuple(map(var -> allocate(var, model), group[ns])))
+                , namespaces
+            )
+        )
+>>>>>>> main
     )
 
     return merge(nt1, nt2)